Banana.Syntax.AST

Description

This module holds the data types that are uses to represent the parsed program.

Documentation

data Type

The possible types in the Banana language

Constructors

Number

Array Integer Type

Void

data VarDecl a

A variable declaration of the form "var x: num"

Constructors

VarDecl

Fields

varName :: a

varType :: Type

data FuncCall a

Function call, with the same representation as for expressions

Constructors

FuncCall

Fields

func :: a

args :: [Expr a]

data Expr a

An arithmetic expression

Constructors

Add (Expr a) (Expr a)
Addition (+)

Sub (Expr a) (Expr a)
Subtraction (-)

Mul (Expr a) (Expr a)
Multiplication (*)

Div (Expr a) (Expr a)
Division (/)

Var a
Variable (identifier)

Lit Double
A literal value

Not (Expr a)
Unary logical not (not)

Or (Expr a) (Expr a)
Logical OR (or)

And (Expr a) (Expr a)
Logical AND (and)

Eq (Expr a) (Expr a)
Equal to (=)

NEq (Expr a) (Expr a)
Not Equal to (/=)

Less (Expr a) (Expr a)
Less than (<)

More (Expr a) (Expr a)
Bigger than (>)

LEq (Expr a) (Expr a)
Less than or Equal to (<=)

MEq (Expr a) (Expr a)
Bigger than or Equal to (>=)

FuncCallExpr (FuncCall a)
A function call as an expression

data VarAssign a

A variable assignment of the form "x := y"

Constructors

VarAssign

Fields

varAssignee :: Expr a

LHS

varExpr :: Expr a

RHS

data IfStatement a

A definition of if statement

Constructors

If

Fields

condition :: [Expr a]

statement :: [Statement a]

data Statement a

A wrapper to collect all statement types together

Constructors

VarDeclaration (VarDecl a)

VarAssignment (VarAssign a)

FuncCallStatement (FuncCall a)

IfStatement (IfStatement a)

data Function a

A function definition

Constructors

Function

Fields

params :: [VarDecl a]

returnType :: Type

body :: [Statement a]

data Program a

The AST node representing the entire program with all of its structures

Constructors

Program

Fields

statements :: [Statement a]

functions :: [Function a]

varDecls :: Program a -> [VarDecl a]

assignments :: Program a -> [VarAssign a]

ifStatements :: Program a -> [IfStatement a]