Banana.Syntax.Parser

Description

The parser gets fed the token generated by the lexer and generates an AST from them representing the program.

Documentation

parseType :: Parser Type

Parse a type

parseVarDecl :: Parser (VarDecl String)

Parse variable declarations

parseExpr :: Parser (Expr String)

Parse an arithmetic expression

parseVarAssign :: Parser (VarAssign String)

Parse the variable assignment statement

parseProgram :: Parser (Program String)

Parse a program consisting of variable declarations and assignments

parseFuncCall :: Parser (FuncCall String)

Parse a function call