Banana.TypeCheck.Checker

Description

Given the parsed AST, the Semantic Checker verifies that the input program is valid and transforms the AST into another intermediate representation for code generation.

Documentation

validLHS :: Expr a -> Bool

Checks if a given expression is a valid LHS in an assignment

varToReference :: Eq a => Expr a -> [VarDecl a] -> Maybe (VarDecl a)

Looks up the variable declaration that declares a given variable

assignmentsAreValid :: Program a -> Bool

Checks if all assignments are valid

varDeclsAreUnique :: Eq a => Program a -> Bool

Checks that no 2 variables are declared with the same name

allVarsDeclared :: Eq a => Program a -> Bool

Checks that all variables used in expressions are declared beforehand

checkAll :: Eq a => Program a -> Bool

Runs all the checkers, combining their output